Add mortgage function, change property dataclass to include ID and level.
Currently, property ID is loosely defined in the Game class and referenced in the Player dataclass. This PR attempts to streamline property ID and at least give it some weight. Constant property IDs are required for house building and mortgaging, because the player needs to access them from anywhere on the board.