Add trade function to player dataclass
Yes, this only takes in one property. This is to handle the lowest edge case of just giving away a property. If you want to do a trade you'd have to call it twice on both players. If we don't want the ability to give away properties then I will change it. I would also consider changing the function name to transfer.